Compilable linux fixes.

+++ common/sim/building.cpp
@@ -295,7 +295,7 @@ void DrawBl()
 		float pitch = 0;
 		float yaw = 0;
 		Matrix modelmat;
-		float radians[] = {DEGTORAD(pitch), DEGTORAD(yaw), 0};
+		float radians[] = {static_cast<float>(DEGTORAD(pitch)), static_cast<float>(DEGTORAD(yaw)), 0};
 		modelmat.setTranslation((const float*)&b->drawpos);
 		Matrix rotation;
 		rotation.setRotationRadians(radians);
